Move escaping functions to their own module
