+++ snuba/replacer.py
@@ -8,12 +8,12 @@ from typing import Any, Mapping, Optional, Sequence
 import simplejson as json
 
 from snuba.clickhouse import DATETIME_FORMAT
+from snuba.clickhouse.escaping import escape_col, escape_string
 from snuba.clickhouse.native import ClickhousePool
 from snuba.datasets.dataset import Dataset
 from snuba.datasets.factory import enforce_table_writer
 from snuba.processor import InvalidMessageType, InvalidMessageVersion, _hashify
 from snuba.redis import redis_client
-from snuba.util import escape_col, escape_string
 from snuba.utils.metrics.backends.abstract import MetricsBackend
 from snuba.utils.streams.batching import AbstractBatchWorker
 from snuba.utils.streams.consumers.backends.kafka import KafkaMessage
